\name{new_parser_def}
\alias{new_parser_def}
\title{
Constructor of ParserDef class
}
\description{
This is a constructor of \code{\link{ParserDef}} class.
}
\usage{
new_parser_def()
}
\value{
  ParserDef S4 class object
}
\seealso{
  \code{\link{ParserDef-class}}
  \code{\link{defineOptions-package}}
}
\examples{
new_parser_def()
}
